move the "delete function" to the "post detail" page
