+++ src/featuers/posts/PostExcerpt.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import moment from "moment";
 import { grey } from "@mui/material/colors";
 import {
@@ -14,20 +14,8 @@ import {
   Typography,
   CardActions,
   Checkbox,
-  Dialog,
-  DialogTitle,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  Button,
 } from "@mui/material";
-import {
-  MoreVert,
-  Star,
-  Delete,
-  ThumbUp,
-  ThumbUpOutlined,
-} from "@mui/icons-material";
+import { MoreVert, Star, ThumbUp, ThumbUpOutlined } from "@mui/icons-material";
 import { useNavigate } from "react-router-dom";
 
 import { useDispatch } from "react-redux";
@@ -37,24 +25,12 @@ import { useDispatch } from "react-redux";
 // 1. if user, update card header info
 
 const PostExcerpt = ({ post, setCurrentId }) => {
-  const [openAlert, setOpenAlert] = useState(false);
   const navigate = useNavigate();
   const dispatch = useDispatch();
   const imageArray = post.image;
 
   const openPost = () => {
     navigate(`/posts/${post._id}`);
-  };
-
-  const clickOpenAlert = () => {
-    setOpenAlert(true);
-  };
-  const closeAlert = () => {
-    setOpenAlert(false);
-  };
-  const handleDelete = () => {
-    // dispatch(deletePost(post._id));
-    console.log("dispatch delete");
   };
 
   const handleLike = () => {
@@ -66,11 +42,8 @@ const PostExcerpt = ({ post, setCurrentId }) => {
       <CardHeader
         avatar={<Avatar sx={{ bgcolor: grey[500] }}>J</Avatar>}
         action={
-          <Tooltip title="Open">
-            <IconButton
-              aria-label="open"
-              onClick={() => setCurrentId(post._id)}
-            >
+          <Tooltip title="More">
+            <IconButton onClick={() => setCurrentId(post._id)}>
               <MoreVert />
             </IconButton>
           </Tooltip>
@@ -117,26 +90,6 @@ const PostExcerpt = ({ post, setCurrentId }) => {
             <Checkbox icon={<Star />} checkedIcon={<Star color="red" />} />
           </IconButton>
         </Tooltip>
-
-        <Tooltip title="delete" arrow>
-          <IconButton aria-label="add to favorites" onClick={clickOpenAlert}>
-            <Delete />
-          </IconButton>
-        </Tooltip>
-        <Dialog open={openAlert} onClose={closeAlert}>
-          <DialogTitle>Move to trash?</DialogTitle>
-          <DialogContent>
-            <DialogContentText>
-              The action can not be reversed. Are you sure to delete the post?
-            </DialogContentText>
-          </DialogContent>
-          <DialogActions>
-            <Button onClick={closeAlert}>Cancel</Button>
-            <Button onClick={handleDelete} color="secondary">
-              Delete
-            </Button>
-          </DialogActions>
-        </Dialog>
       </CardActions>
     </Card>
   );
